It compiles
