+++ src/component/activity_list/mod.rs
@@ -1,30 +1,149 @@
-pub mod list;
 pub mod chart;
-pub mod sort_dialog;
+pub mod list;
 pub mod rank_dialog;
+pub mod sort_dialog;
 
+use crossterm::event::Event;
 use tui::{
-    backend::Backend,
     layout::{Constraint, Layout},
-    Frame, widgets::TableState,
+    prelude::Buffer,
+    style::Style,
+    widgets::{Block, Borders, Clear, Paragraph, StatefulWidget, Table, TableState, Widget},
 };
+use tui_input::backend::crossterm::EventHandler;
 use tui_input::Input;
 
-use crate::{app::App, event::keymap::MappedKey};
+use crate::{
+    app::App,
+    event::{
+        input::InputEvent,
+        keymap::{MappedKey, StravaEvent},
+    },
+    store::activity::SortOrder,
+    ui::{centered_rect_absolute, color::ColorTheme},
+};
 
-pub fn handle(app: &mut App, key: MappedKey) {
-    list::handle(app, key)
-}
-pub fn draw<B: Backend>(
-    app: &mut App,
-    f: &mut Frame<B>,
-    area: tui::layout::Rect,
-) -> Result<(), anyhow::Error> {
-    let rows =
-        Layout::default().constraints(vec![Constraint::Percentage(50), Constraint::Percentage(50)]).split(area);
-    chart::draw(app, f, rows[1])?;
-    list::draw(app, f, rows[0])?;
-    Ok(())
+use self::list::activity_list_table;
+
+use super::{table_status_select_current, View};
+
+pub struct ActivityList {}
+
+impl View for ActivityList {
+    fn handle(&self, app: &mut App, key: MappedKey) {
+        if app.activity_list.filter_dialog {
+            let matched = match key.strava_event {
+                StravaEvent::Enter => {
+                    app.filters.filter = app.activity_list.filter_text_area.value().to_string();
+                    app.activity_list.filter_dialog = false;
+                    app.activity_list.table_state().select(Some(0));
+                    app.send(InputEvent::Reload);
+                    true
+                }
+                _ => false,
+            };
+            if matched {
+                return;
+            }
+
+            app.activity_list
+                .filter_text_area
+                .handle_event(&Event::Key(key.key_event));
+            return;
+        }
+
+        if app.activity_list.sort_dialog {
+            sort_dialog::handle(app, key);
+
+            return;
+        }
+        if app.activity_list.rank_dialog {
+            rank_dialog::handle(app, key);
+
+            return;
+        }
+        match key.strava_event {
+            StravaEvent::Quit => app.quit = true,
+            StravaEvent::ToggleUnitSystem => {
+                app.unit_formatter = app.unit_formatter.toggle();
+            }
+            StravaEvent::ToggleSortOrder => {
+                app.filters.sort_order = match app.filters.sort_order {
+                    SortOrder::Asc => SortOrder::Desc,
+                    SortOrder::Desc => SortOrder::Asc,
+                };
+                app.send(InputEvent::Reload);
+            }
+            StravaEvent::Down => app.next_activity(),
+            StravaEvent::Up => app.previous_activity(),
+            StravaEvent::Filter => toggle_filter(app),
+            StravaEvent::Sort => toggle_sort(app),
+            StravaEvent::Rank => toggle_rank(app),
+            StravaEvent::Enter => table_status_select_current(app),
+            StravaEvent::Refresh => app.send(InputEvent::Sync),
+            StravaEvent::IncreaseTolerance => {
+                app.filters.anchor_tolerance_add(0.01);
+                app.send(InputEvent::Reload)
+            }
+            StravaEvent::DecreaseTolerance => {
+                app.filters.anchor_tolerance_add(-0.01);
+                app.send(InputEvent::Reload);
+            }
+            StravaEvent::Anchor => {
+                app.anchor_selected();
+                app.send(InputEvent::Reload);
+            }
+            _ => (),
+        }
+    }
+
+    fn draw(&self, app: &mut App, f: &mut Buffer, area: tui::layout::Rect) {
+        let rows = Layout::default()
+            .constraints(vec![Constraint::Percentage(50), Constraint::Percentage(50)])
+            .split(area);
+        chart::draw(app, f, rows[1]);
+
+        let activities = &app.activities();
+
+        if app.activity_list.table_state().selected().is_none() && !activities.is_empty() {
+            app.activity_list.table_state().select(Some(0));
+        }
+
+        let table = activity_list_table(app, activities);
+        <Table as StatefulWidget>::render(table, rows[0], f, app.activity_list.table_state());
+
+        if app.activity_list.filter_dialog {
+            let rect = centered_rect_absolute(64, 3, area);
+            let p = Paragraph::new(app.activity_list.filter_text_area.value()).block(
+                Block::default()
+                    .borders(Borders::ALL)
+                    .title("Filter")
+                    .border_style(Style::default().fg(ColorTheme::Dialog.to_color())),
+            );
+
+            // TODO: support cursor pos
+            //f.set_cursor(
+            //    1 + rect.x + app.activity_list.filter_text_area.visual_cursor() as u16,
+            //    rect.y + 1,
+            //);
+
+            Clear.render(area, f);
+            p.render(area, f);
+
+            return;
+        }
+
+        if app.activity_list.sort_dialog {
+            sort_dialog::draw(app, f, area);
+
+            return;
+        }
+        if app.activity_list.rank_dialog {
+            rank_dialog::draw(app, f, area);
+
+            return;
+        }
+    }
 }
 
 pub enum ActivityListMode {
@@ -53,11 +172,20 @@ impl ActivityViewState {
 }
 
 impl ActivityListState {
-    pub fn table_state(&mut self) -> &mut TableState
-    {
+    pub fn table_state(&mut self) -> &mut TableState {
         match self.mode {
             ActivityListMode::Normal => &mut self.table_state,
             ActivityListMode::Anchored => &mut self.anchored_table_state,
         }
     }
+}
+fn toggle_filter(app: &mut App) {
+    app.activity_list.filter_dialog = !app.activity_list.filter_dialog;
+}
+
+fn toggle_sort(app: &mut App) {
+    app.activity_list.sort_dialog = !app.activity_list.sort_dialog;
+}
+fn toggle_rank(app: &mut App) {
+    app.activity_list.rank_dialog = !app.activity_list.rank_dialog;
 }
